+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `Dependency` now checks against modified values instead of actual values when checking numeric properties.
+- `GameObject.addChild()` now checks whether an invalid child is being added.
 
 ### Refactored